Surgery for getting this to work in OS X. EOM

+++ deploy.bash
@@ -1,17 +1,25 @@
 #!/bin/bash
 
 if [ -f ~/.bashrc ]; then
-    rm ~/.bashrc
+  rm ~/.bashrc
 fi
 
 if [ -f ~/.bash_profile ]; then
-    rm ~/.bash_profile
+  rm ~/.bash_profile
 fi
 
 if [ -f ~/.inputrc ]; then
-    rm ~/.inputrc
+  rm ~/.inputrc
 fi
 
-cp bashrc ~/.bashrc
-cp bash_profile ~/.bash_profile
-cp inputrc ~/.inputrc
+if [ -f bashrc ]; then
+  cp bashrc ~/.bashrc
+fi
+
+if [ -f bash_profile ]; then
+  cp bash_profile ~/.bash_profile
+fi
+
+if [ -f inputrc ]; then
+  cp inputrc ~/.inputrc
+fi
